Move DockerCredentialHelper.SYSTEM_CREDENTIAL_HELPER_DELEGATE to own class
to make it easier to test
